pycam/pyproject.toml
[tool.poetry]
name = "pycam"
version = "0.1.0"
description = ""
authors = ["Craigryy <harria.hj@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.0"
graphviz = "^0.20.0"
django = "^5.0.0"
django-allauth = "^65.0.0"
pillow = "^10.0.0"
django-pwa = "^1.1.0"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.9.9"
PyJWT = "^2.8.0"
cloudinary = "^1.44.0"
numpy = "^2.2.5"
uvicorn = "^0.34.2"
dj-database-url = "^2.3.0"
flake8 = "^7.2.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
